requirements.txt
Flask
pymongo
python-dotenv
pandas
langchain
langgraph
sentence-transformers
werkzeug
tavily-python
google-search-results
langchain_openai
langchain_core
langchain-community
seo-keyword-research-tool
langchain-google-genai
openai
openpyxl
transformers
langchain-text-splitters
ragas==0.2.7
datasets
